RalcoIndia (in my git copy)/realcoindia/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.2
Django==4.0.5
django-braces==1.15.0
django-cors-headers==3.13.0
django-rest-swagger==2.2.0
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.0
idna==3.3
itypes==1.2.0
Jinja2==3.1.2
jwt==1.3.1
MarkupSafe==2.1.1
openapi-codec==1.3.2
pycparser==2.21
PyJWT==2.4.0
pytz==2022.1
PyYAML==6.0
requests==2.28.0
simplejson==3.17.6
sqlparse==0.4.2
twilio==7.9.3
uritemplate==4.1.1
urllib3==1.26.9
